Bugfix: Rasterplot sorting
Bug happened due to the inability to save changes in a Mcs object,
which is troubling when we want to sort the arrays to make up for
different channel orders.

The solution is a hack where we copy the entire rasterplot code from
the Msc package and replace the timestamp object with a normal
cell array. Hopefully this will work...